+++ examples/next-langchain/app/api/chat/route.ts
@@ -1,29 +1,29 @@
-import { StreamingTextResponse, LangChainStream, Message } from 'ai';
-import { ChatOpenAI } from 'langchain/chat_models/openai';
+import { ChatOpenAI } from '@langchain/openai';
+import { LangChainAdapter, Message, StreamingTextResponse } from 'ai';
 import { AIMessage, HumanMessage } from 'langchain/schema';
 
 export const dynamic = 'force-dynamic';
+export const maxDuration = 60;
 
 export async function POST(req: Request) {
-  const { messages } = await req.json();
+  const {
+    messages,
+  }: {
+    messages: Message[];
+  } = await req.json();
 
-  const { stream, handlers } = LangChainStream();
-
-  const llm = new ChatOpenAI({
-    streaming: true,
+  const model = new ChatOpenAI({
+    model: 'gpt-3.5-turbo-0125',
+    temperature: 0,
   });
 
-  llm
-    .call(
-      (messages as Message[]).map(m =>
-        m.role == 'user'
-          ? new HumanMessage(m.content)
-          : new AIMessage(m.content),
-      ),
-      {},
-      [handlers],
-    )
-    .catch(console.error);
+  const stream = await model.stream(
+    messages.map(message =>
+      message.role == 'user'
+        ? new HumanMessage(message.content)
+        : new AIMessage(message.content),
+    ),
+  );
 
-  return new StreamingTextResponse(stream);
+  return new StreamingTextResponse(LangChainAdapter.toAIStream(stream));
 }