+++ packages/core/core/generate-text/stream-text.test.ts
@@ -6,6 +6,7 @@ import { convertReadableStreamToArray } from '../test/convert-readable-stream-to
 import { MockLanguageModelV1 } from '../test/mock-language-model-v1';
 import { createMockServerResponse } from '../test/mock-server-response';
 import { streamText } from './stream-text';
+import { formatStreamPart } from '../../streams';
 
 describe('result.textStream', () => {
   it('should send text deltas', async () => {
@@ -264,6 +265,134 @@ describe('result.toAIStream', () => {
         result.toAIStream().pipeThrough(new TextDecoderStream()),
       ),
       ['0:"Hello"\n', '0:", "\n', '0:"world!"\n'],
+    );
+  });
+
+  it('should invoke callback', async () => {
+    const result = await streamText({
+      model: new MockLanguageModelV1({
+        doStream: async ({ prompt, mode }) => {
+          return {
+            stream: convertArrayToReadableStream([
+              { type: 'text-delta', textDelta: 'Hello' },
+              { type: 'text-delta', textDelta: ', ' },
+              { type: 'text-delta', textDelta: 'world!' },
+            ]),
+            rawCall: { rawPrompt: 'prompt', rawSettings: {} },
+          };
+        },
+      }),
+      prompt: 'test-input',
+    });
+
+    const events: string[] = [];
+
+    await convertReadableStreamToArray(
+      result
+        .toAIStream({
+          onStart() {
+            events.push('start');
+          },
+          onToken(token) {
+            events.push(`token:${token}`);
+          },
+          onText(text) {
+            events.push(`text:${text}`);
+          },
+          onCompletion(completion) {
+            events.push(`completion:${completion}`);
+          },
+          onFinal(completion) {
+            events.push(`final:${completion}`);
+          },
+        })
+        .pipeThrough(new TextDecoderStream()),
+    );
+
+    assert.deepStrictEqual(events, [
+      'start',
+      'token:Hello',
+      'text:Hello',
+      'token:, ',
+      'text:, ',
+      'token:world!',
+      'text:world!',
+      'completion:Hello, world!',
+      'final:Hello, world!',
+    ]);
+  });
+
+  it('should send tool call and tool result stream parts', async () => {
+    const result = await streamText({
+      model: new MockLanguageModelV1({
+        doStream: async ({ prompt, mode }) => {
+          assert.deepStrictEqual(mode, {
+            type: 'regular',
+            tools: [
+              {
+                type: 'function',
+                name: 'tool1',
+                description: undefined,
+                parameters: {
+                  $schema: 'http://json-schema.org/draft-07/schema#',
+                  additionalProperties: false,
+                  properties: { value: { type: 'string' } },
+                  required: ['value'],
+                  type: 'object',
+                },
+              },
+            ],
+          });
+          assert.deepStrictEqual(prompt, [
+            { role: 'user', content: [{ type: 'text', text: 'test-input' }] },
+          ]);
+
+          return {
+            stream: convertArrayToReadableStream([
+              {
+                type: 'tool-call',
+                toolCallType: 'function',
+                toolCallId: 'call-1',
+                toolName: 'tool1',
+                args: `{ "value": "value" }`,
+              },
+              {
+                type: 'finish',
+                finishReason: 'stop',
+                logprobs: undefined,
+                usage: { completionTokens: 10, promptTokens: 3 },
+              },
+            ]),
+            rawCall: { rawPrompt: 'prompt', rawSettings: {} },
+          };
+        },
+      }),
+      tools: {
+        tool1: {
+          parameters: z.object({ value: z.string() }),
+          execute: async ({ value }) => `${value}-result`,
+        },
+      },
+      prompt: 'test-input',
+    });
+
+    assert.deepStrictEqual(
+      await convertReadableStreamToArray(
+        result.toAIStream().pipeThrough(new TextDecoderStream()),
+      ),
+      [
+        formatStreamPart('tool_call', {
+          toolCallId: 'call-1',
+          toolName: 'tool1',
+          args: { value: 'value' },
+        }),
+        formatStreamPart('tool_result', {
+          toolCallId: 'call-1',
+          toolName: 'tool1',
+          args: { value: 'value' },
+          result: 'value-result',
+        }),
+      ],
     );
   });
 });