+++ packages/openai/src/openai-chat-settings.ts
@@ -1,5 +1,7 @@
 // https://platform.openai.com/docs/models
 export type OpenAIChatModelId =
+  | 'gpt-4o'
+  | 'gpt-4o-2024-05-13'
   | 'gpt-4-turbo'
   | 'gpt-4-turbo-2024-04-09'
   | 'gpt-4-turbo-preview'