+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/anthropic
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [325ca55]
+  - @ai-sdk/provider@0.0.4
+  - @ai-sdk/provider-utils@0.0.7
+
+## 0.0.11
+
+### Patch Changes
+
+- 5b01c13: feat (ai/core): add system message support in messages list
+
 ## 0.0.10
 
 ### Patch Changes